api/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
pymongo==4.7.3
gunicorn==22.0.0
